Add the SuperCold / "lower" CachePriority into the ESE IO trace.
Andrew asked a very reasonable question on esedev, if our BFAvail IOs are related to super cold or cache priority.  Very easy to add the SuperCold / lower CachePriority into the ESE IO trace's IORF / IOReasonFlags.

The actual flag already existed but was not used anywhere.  I don't know if I just forgot to complete this piece of trace context, or if someone broke it / removed it at some point.

[Substrate:6ffc195b608c8afe9d4a07b8c54112a8b32e7f13]