ESE should recognize the real volume hosting a file indirected through a symbolic link
The LogsOnSSD feature moves log files to different storage tiers and leaves behind symbolic links to make the set of log files seem contiguous.  ESE isn't aware of this and thinks that such a log file is on the volume holding the sym link rather than its real host volume.

This change makes ESE aware of the real host volume of the symlink.  This is done by modifying ErrPathRoot to detect files that are symlinks and then do the extra steps to find the true underlying path of the file.  This will allow us to get the correct volume for the file which will allow us to get all the correct corresponding properties for the file.

Ultimately the motivation is to get correct ESEIO telemetry including if the IO is truly being done to HDD or flash storage.  This will allow us to see the actual real HDD IO savings for LogsOnSSD.

[Substrate:766cf3c92f0ad475ed21551d2e3a90a059e861f1]