Save 1% HDD IOPS caused by incorrect nullification of preread for LRPAGEMOVE
We recently noticed that we were touching pages that were not preread during log replay in Substrate.  We added some more telemetry and mysteriously almost all of it was caused by replay of LRPageMove operations which are generated mostly by OLD2.  This accounts for 1% of our HDD IOPS.

It turns out that OLD2 repeatedly emits page move operations in a pattern where we reference a page (e.g. the right page) and then later this page is referenced by a LRNEWPAGE or another LRPAGEMOVE as a destination page.  This later reference causes us to attempt to suppress the preread because we don't want to preread new page operations.  However, this override is incorrect because we clearly do need to read that page earlier in the log stream.

The fix is to only suppress a page reference for preread if it is for a new page operation and we haven't already accessed that page in the current preread domain (currently a single log file).

This change contains a test case that explicitly covers this condition.

[Substrate:3b64b0a5de273f9bacdd264d57cedb3a2ce57e8e]