Distinguish replay of logs in the required range vs new logs in ESE IO telemetry
We currently start out as iortRecovery and then switch to iortRecoveryRedo when replay of logs begins.  If we start to replay a log file when all databases are no longer in their required ranges then we will now switch to iortRecoveryRedoNewLogs to indicate that.

The goal here is to measure how much IO we are "wasting" in the service by replaying logs that have already been played once.  This happens for example on a replica restart.  This should be rare but we want to measure it.

[Substrate:d6dd9038e2e5bdc587be5d55341095f1fafa01f1]