Enhance ESE IO Telemetry to identify avoidable recovery reads
We will tag recovery reads that could be avoided by using the page image in the log with iorfFill.

This existing iorf is only used by writes to tag log file pattern files.  This is easier than defining a new one because these flags are very limited in number and because this instrumentation is temporary.

[Substrate:caec181e691a9346b2dbb6c6da0a8eed7e2fd033]