Avoid iorpDatabaseShrink leakage into BF dirty traces
We have a test that asserts that BF-dirty traces must not have an IORP (primary reason) set. That test fails every now and then with an improper iorpDatabaseShrink set. This change limits iorpDatabaseShrink to non-BF functions (where an IORP is required), thus preventing its leakage into BF and potentially BF-dirty.

[Substrate:c2c912b2de2033d5f4cb639f60b32f6388866ddf]