Enhance ESE IO Telemetry to identify avoidable recovery reads (fix)
We will tag recovery reads that could be avoided by using the page image in the log with iorfFill.

This existing iorf is only used by writes to tag log file pattern files. This is easier than defining a new one because these flags are very limited in number and because this instrumentation is temporary.

Fix:  actually set the IOREASONFLAGS we carefully saved so that it makes it to the telemetry

[Substrate:ab5580e9274a63c609bb60a0328fc069e0b82b0d]