sync: Reimplement CSemaphore for better performance
The existing implementation uses a spinlock that falls back to a kernel
semaphore.  When the contention is high and cannot be alleviated by
the spinlock, this approach fails to deliver an adequate throughput.
The probable cause for that is the general slowness of the semaphore
object created with `CreateSemaphore()`.

The new implementation is based on the `WaitOnAddress()` primitive and
delivers an up to 20x synthetic improvement.  Since the CSemaphore is
used as a building block for other synchronization primitives, such
as CCriticalSection, this translates into the up to ~5.4x throughput
improvement in a benchmark with multiple concurrent db writers.

The benchmarks below were conducted in an environment with an i9-9900K
CPU and an NVMe SSD.

Synthetic benchmark based on the adjusted semaphoreperf.cpp test:

    1 thread:    93850 KOps/Thread  →  93950 KOps/Thread
    2 threads:   20445 KOps/Thread  →  30165 KOps/Thread  (+48 %)
    3 threads:    6083 KOps/Thread  →  10897 KOps/Thread  (+79 %)
    8 threads:     102 KOps/Thread  →   2058 KOps/Thread  (+1917 %)
   16 threads:      45 KOps/Thread  →    719 KOps/Thread  (+1498 %)

Real benchmark with multiple concurrent db writers:

  (The numbers represent the amount of complete "write sequences"
   and are meaningful only as a relative measurement of throughput)

    1 writer:     8443 Sequences  →   8562 Sequences
    2 writers:   14934 Sequences  →  15331 Sequences
    3 writers:   17434 Sequences  →  17386 Sequences
    4 writers:   16583 Sequences  →  17997 Sequences
    5 writers:   15924 Sequences  →  17431 Sequences
    6 writers:    8437 Sequences  →  17223 Sequences  (+104 %)
    7 writers:    5020 Sequences  →  18145 Sequences  (+261 %)
    8 writers:    4131 Sequences  →  17670 Sequences  (+339 %)
    9 writers:    4168 Sequences  →  17438 Sequences  (+318 %)
    10 writers:   3670 Sequences  →  19709 Sequences  (+437 %)

On Win64, the size of CSemaphore remains unchanged (8 bytes).

On Win32, the size of CSemaphore changes from 4 to 8 bytes, so the
patch updates the size assertions for structures like PIB.