[Cleanup][ESE-Linux] Replace IN, __in, and OUT, __out annotations with _In_ and _Out_
This change includes many files, but it's tediously simple.

- There is a comment in `cc.hxx` that `IN` and `OUT` should be replaced with the [newer Windows 8+ SAL annotations](https://docs.microsoft.com/en-us/previous-versions/visualstudio/visual-studio-2013/ms182032(v=vs.120))  `_In_` and `_Out_`
- `__in` and `__out` are [older-style header annotations](https://docs.microsoft.com/en-us/windows/win32/winprog/header-annotations) (according to the docs, pre-Windows 8) that conflict with attributes in headers on Linux. They can be replaced with `_In_` and `_Out_`

Therefore, this change is:
- `IN` and `__in` -> `_In_`
- `OUT` and `__out` -> `_Out_`

I tried to "ban" `__in` and `__out`, but, as I say in the comment, a small handful of these crept back into some Windows headers that we include. Interestingly, the same Windows headers we keep under `XXXX` are older and only have new-style annotations.

[Substrate:acde71839e137b226de3b8b8cf9d3f831248744b]