Change OLD2 to require append/hotpoint sequential space hints for eligibility
OLD2 tries to use the continuous space pool to allocate space for defragmentation even if the table is not setup to use this type of space by default.

Unfortunately, this does not result in the desired effects.  We have seen in test and in production that OLD2 is creating trees that are still substantially fragmented.  We see single passes of OLD2 defrag much more than expected.  We have seen OLD2 defrag create trees that are even more fragmented than they originally were.

This was hit in production by the store's lazy indices which have for a very long time been marked for optimization for sequential scan but were not setup for sequential append (because they are not).  This is causing OLD2 to not converge and to repeatedly defrag some lazy indices.  This causes acute problems for some databases (single copy due to high copy queue).  This is overall consuming 2% of our HDD IOPS just for "defragging" lazy indices.

We now require that a table be configured for sequential appends to be eligible for OLD2.  This will effectively disable OLD2 for lazy indices in production.

[Substrate:f84f2e3d432e3d392ed4ad983bec2f64e3e5315b]