Use LOSStrCompare, not wcscmp
More forcing of use of string support routines rather than native APIs.

This is in multiple steps.  The first step is to note that the existing LOSStrCompare is odd.  It sorts longer strings as greater than shorter strings, regardless of characters.  So, the following is the existing sort order:

"zoo" < "apes" < "zebra"

So, I've renamed the API to get the compiler to take me to all uses, and we find that all but one only care about 0 or !0 (where 0 means same length, same characters).  Tweaked the one odd user and changed the actual API to do a more common comparison.

Next step was un-renaming and finding/changing wcscmp() users.

And final step was to polish the compare routines to be as performant as possible.

[Substrate:fe134bd28652e7d63f83abc0ab6151b873f0ee54]