sync: Block stealing the semaphore from woken threads
Allowing to steal the semaphore from the threads that are woken up
introduces a fairness issue and can lead to starvations.

The underlying `WakeOnAddress()` primitive by itself allows for a fair
implementation, because it guarantees that the threads are woken up in
FIFO order [1].  However, if we don't block the other threads from
stealing the semaphore from the waiters right after waking them up,
those waiters may keep getting starved out.

This behavior is shown by the new regression test, CSemaphoreFairnessTest.

[1] https://docs.microsoft.com/windows/win32/api/synchapi/nf-synchapi-wakebyaddresssingle