sync: Rewrite the semaphore wait loop
Simplify the code by switching to a single loop, instead of using a
nested loop for the OS-level wait.  Ensure that when returning from
the OS-level wait, we try to acquire the semaphore as soon as possible,
without any intermediate actions such as adjusting the remaining time —
as this is better from the standpoint of working under contention.