cpage_test.cxx: Fix build error caused by mismatching defines
The PageValidationBig() test uses #ifdef DEBUG and #ifndef RTM as if they
were interchangeable, but this is not so at least in the RelWithDebInfo
cmake configuration.

This currently causes the following errors in that configuration:

   error C2065: 'fPreviouslySet': undeclared identifier

Fix this by consistently using #ifndef RTM, as it's already being used
in the other similar tests in cpage_test.cxx.

(Perhaps, the `RelWithDebInfo` configuration could use the same defines
 as `Release`, but the code discrepancy seems to be worth fixing by itself.)