Fix a potential stack corruption in CResource::ErrGetParam()
Current implementation always does *pdwParam = 0, where pdwParam is a
`DWORD_PTR`.  On x64, this translates into a 8-byte write to the memory
location pointed by pdwParam.

However, for the JET_resoperTag param, this location may point to a
buffer of a lesser size.  For example, CResource::PvAlloc_() passes a
buffer of size `JET_resTagSize + 1` (= 5).  The remaining part of the
8-byte write happens into a memory location after the allocated buffer.
Potentially this can either corrupt the stack or cause a crash.

Fix this by making the CResource::ErrGetParam() specially handle the
JET_resoperTag parameter, as it is already done in the similar method,
CResourceManager::ErrGetParam().