Merged PR 1273485: Improve IO efficiency of database scrubbing especially during replay
Our recent improvements to EseIO telemetry have shown that we are currently using about 1.25% of our HDD IOPS in SDFv2 doing sync reads to replay scrub operations.  Where enabled, scrub is performed concurrently to database scan so this IO should not happen at all.

We think that this is caused by a combination of scan check replay marking pages as super cold immediately and other cache activity causing these pages to be evicted before the scrub can be replayed.

This change does the following:
 - The do time database scan code is improved to mark all pages as super cold at once rather than as they are processed.  This should improve the odds that scrubbed pages are written as one IO rather than many smaller ones.
 - The redo time database scan code is improved to mark all pages as super cold at the end of each log file rather than as they are processed.  This should improve the odds that scrubbed pages will hit the cache.  This should also improve the odds that scrubbed pages are written as one IO rather than many smaller ones.

This should reduce IO substantially in environments where scrubbing is enabled.

[Substrate:7104d18d7260e34674e7a793482094ef1969b804]