OSSTEST: Drop the all tests >= ~4 seconds (happens to be top 10 most expensive tests) temporarily.
We have many many changes to sync for the last 4 years of catch-up and the test-script loop body is quite a factor in getting this done.  So this is a temporary necessity to accelerate GitHub syncing.  We will remove this when 'caught up' in some fashion.

Reduces OSS test pass from 5:02 minutes to ~:39 seconds.

Note: This change is back-dated to fit within the order of the filled in historical commits.