Avoid repeating database scan checks during recovery
We found in prod via EseIO telemetry that we are replaying the same database scan checks during recovery in some cases.  This happens when a passive is restarted and we are playing through the required range of the logs.  These scan checks have already occurred so they don't need to be repeated.

This is consuming 0.18% of HDD IOPS in the WW ring.  This is a significant percentage of the HDD IOPS we "waste" replaying logs that have already been played once.

DbDivergenceScanSuite.RepeatedlyScanLogRecordsOnCrash was modified to directly test this modification.

[Substrate:1e4ac180a18adae91a4824a70d6e0427fd6ce0a4]