Restore ESE database space cache
This cache was inadvertently deleted when another space caching feature was recently added.  This caused a CPU regression of 0.15% or so in prod.

This change is a simple revert of the deleted caching code.  Almost no modification was required.

A test case was added to cover this functionality which was never covered at all.  The test verifies that no database pages are touched when accessing the cached information.

[Substrate:eaeedc47a5df32412e1a1cf888324cce79116181]