Some reordering in ExtentPageCountCache
Move CATIPossiblyResetUpdatingExtentPageCountCacheFlag to after we may be doing ErrFILECloseTable() to avoid inaccurate asserts.

Also, change the BYTE column to actually be a flag, and introduce a new state.  Avoids IsamDelete in ExtentPageCountCache update paths to avoid doing anything that might affect a space tree and need a lock on the pgnoSystemRoot.

[Substrate:b8da0f3940555abb27fa2dfd9108770c248122b1]