ESE Block Cache:  TLS infrastructure
EBC needs thread local storage to track IO requests.

This will be used in two places.  TFileFilter will use this to track IO requests to implement features such as qosIOOptimizeCombinable and qosIODispatchBackground.  THashedLRUKCache will use this to track IO requests to implement async cache IO and to allow coalescing IO requests for cache misses so that they may be combined.

This code is tested by invoking it in TPassThroughCache even though it isn't needed there.

Misc cleanup related to memory management was also done.

This is being spun out of my main branch to manage the size of that review when it happens.

[Substrate:b58e1d9041afd1c56114aa5c0fad0a3cc756ca6f]