Drop 3: comments (#19)
This 3rd iteration completes the initial release of ESE's source code. The content of this drop is the comments. What's not included are things like mentions of internal tools, bug IDs, employees' names, etc.

These comments should help in understanding and contributing to the code.

The goal going forward is to get to the point where we can replay internal commits from the initial drop until the present and into the future, in an automated fashion.