Always write database header when re-attaching it during redo
There is code in the DB-attachment during redo that tries to suppress writing the DB header if it is deemed that no relevant changes have occurred. This could have the side effect of not re-establishing synchronization between the DB and the flush map files in such a way that a mismatched/stale flush map could be incorrectly used in the future.

This code has existed for as long we could see (at least 2004), but the recent changes to suppress DB header writes unless there is a change in the required range made it more likely that changes could be made to the database and a mismatched flush map could be attached, leading to false positive lost flush detections.

This change simply removes that optimization, to make sure DB and flush map headers are always in sync upon re-attaching a DB during redo.

[Substrate:b42caad652d56e695a6aca8037ec290e9013b150]