Skip capturing preimages for deleted tables to avoid capturing 10s of GBs of pages unnecessarily
1.	When we revert back to the time when table might have still existed, the catalog page is going to be reverted and is still going to have the table entries. So, we need to redelete the table and block any other operations.
a) So to redelete after revert, we capture the preimage of table root pages and the space tree pages as part of delete operation.
b) When we revert the table root page which was a non-revertable delete, we will mark the table root page with a special flag (fPageFDPDelete) to indicate that the table needs to be deleted.
c) DbScan will redelete this table when it gets to the table root page which is marked with this special flag.
d) Also, to make it easier to block operations on the secondary tables belonging to this table (secondary index, LV table), we will also capture the preimage of root page of the LV tree and secondary index tree and also mark them with the special flag fPageFDPDelete.
e) Any updates to the tree marked with the special flag will throw an error when the root page is accessed.

2.	Even though a table/tree is disconnected, it is possible there are some updates on the root pages of the table for some cleanup operations and we don't want to overwrite the special flag fPageFDPDelete.
a) To overcome this issue, while reverting root pages, we will read the disk/in-memory page and see if that flag is set and make sure to set that flag in the reverted page we are writing.
b) If the preimage already has the special flag set, but if earlier in the snapshot we captured another preimage of the page, we will re-read the page (either from disk or cache) and just set the special flag on it and not capture the current preimage.
c) While reverting space tree preimages, we don't do anything special. There are treated like any other page preimage.

3.	Another issue is that even though store deletes a disconnected mailbox ~30days after migration, there is no guarantee. It is possible a table which was just created is being deleted or it is possible the root page of the table was recently moved by shrink. In such a case, even the above step wouldn't help us since the catalog will be pointing to a page before the shrink after we revert but the root page we captured would be a different page number.
a) To avoid this issue, we will now add to the catalog a column for the tables to indicate the last time the root page was set.
b) For existing tables which don't have this set, this time will be set the first time they are accessed.
c) This time is only updated by dbshrink during root page move.
d) We will block any non-revertable table deletes if it was too recent (indicated by our lag time = 7days)

4.	One more issue which we came across was normal page updates before table delete. If its within the required range, there is no guarantee if the page was already flushed and if the pa...

[Substrate:98086d044a636c4183da7914988c353eb698c87e]