Use simpler "double" based math for fractional positioning.
Using doubles to hold the fraction removes the need for a 64 bit MulDiv when we switch fractional positioning to 64 bit.

Also, testing the old algorithm with 100,000 randomly inserted rows resulted in different positioning for setting to fractional positions (11900 / 100,000), (5950 / 50,000), (238 / 2,000), and (119 / 1,000), which are all the same fraction while this positions on the same row for each (and it's one of the rows that the old algorithm hits).  The old algorithm suffers from small but real truncation problems when the fraction passed in is the smaller numbers.

[Substrate:3a3fe9e03a6c0a7ae1841458c82a56fd37e7d3a0]