Do ReplaceInIndex for tuples the same way as all the other indices
Do tuples the same way as all the other indices.

I had exempted them at first because I was concerned about the current algorithm working on large data sets, but by now AD has been using this on large multivalued (100s of keys) data sets, so it's time to simplify this.  It also simplifies a different project I'm working on.

[Substrate:35e906885986ad056f5e431b4af96a999554f91c]