Rework some index entry code
I'm working on another CR that introduces an index that is temporarily corrupt on purpose and gets populated over time.  To make that work, I need better control of the insert/delete/update code for 2ndary indices so I can better control when it complains about corruption.  This CR does that.  The other CR that really needs it will come later.

[Substrate:40ec96c3e56cfc88e41fc21ed62a2c555d56355c]