Make perf. counter refresh frequency configurable
Andrew found via CPR that our perfmon thread is #3 in CPU exclusive time on BE servers. Passive Store worker processes spend up to 12% in that function. This fix makes the refresh interval of our performance counters configurable, sets the engine default to 100 msec (which is, on average, how it behaves today) and sets the Service default to 1 sec. Some ESE-level tests set to 0 so that results wouldn't be skewed, as they relied on the model of "max of 10 samples per second", instead of a fixed interval between samples (100 msec now).

[Substrate:a07bde17b97472c79c65cc5bfb8c9248ee552e0a]