DB growth should not violate the configured max DB size to fulfill the preferred growth size
The current code might use a larger size (preferred size) compared to the minimum necessary to fulfill the request if there are shelved extents in the region. Using the preferred size instead of the minimum size might lead to a slight overgrowth beyond the max size.

This fix uses the lesser of the two values when updating the relevant variable so that we won't overgrow unless absolutely necessary (to prevent space leaks).

This was observed in 10 out of 8k+ MCDBs that were shrunk down to their max sizes in the last two months.

[Substrate:a662225fbbec87965f0384a57b88f456298ac10e]