Fix crash in ErrBFFlush in JetCreateIndex
This very old code uses a hack to verify that all pages belonging to a given objid (of an index) are truly flushed before ErrBFFlush returns.  It does this by touching the page buffer.  This page buffer could go missing suddenly which can result in a crash.

The fix ignores all error and clean BFs first which greatly reduces the probability of hitting the crash while preserving the defensive check.

[Substrate:646b0e644ea10d75644bc295c233a887de9e7453]