Reserve a chunk of the ESE error space for users.
Back in AD, I had some stacks that worked primarily in JET error code returns.  However, there were a handful of local conditions I had to check and return errors.  At the time, I looked in the ESE header for a few unused numbers.  This formalizes that into a set of numbers with a promise so end users can slip their own errors into code that primarily deals with JET.

This also does a tweak of an AssertRTL into an AssertTrack for discussion/check-in purposes.

[Substrate:e76fcae76d97c1d86ea8fd5e268bbb2af2c97e30]