Disable Shrink to prevent space leakage
If we're about to quit Shrink due to lack of space at lower offsets, disable it to free up any space which is pending release. This might happen if we succeeded in moving OE split buffers (thus returning the replace buffer for release), but subsequently fail to refill AE split buffers because we ran out of space in lower offsets.

I've never seen this "in the wild", but it hits ~1x/month in the test pass.

[Substrate:cef303ea7fbc21e5cab02190394e8762e328b753]