ESE Block Cache:  Hashed LRUK Cache minimum viable product
This change implements the minimum viable Hashed LRUK Cache.

This cache is NOT checked in enabled.  We would like to do some more work before we start rolling out this feature.

This change contains a huge amount of churn to the EBC testing infrastructure (BlockCache.Interop).  While the curious may find certain aspects of it interesting, a review of those files will not provide much value.  However, the interop layer is used by the ExchangeDatabaseManager to manage the configuration of EBC in production.

[Substrate:8e6a5858ce7d483bb84ad4fafefe9da72116f9cb]