ESE Block Cache:  fix hang related to managing CIOComplete validity table
We use a CMeteredSection to keep a CIOComplete alive between the time we look it up in the table and the time we are done using it.  ~CIOComplete must wait for the users to leave before it can complete the delete.  This was previously done while holding the lock on the table and this caused a deadlock in rare cases.  It is not necessary to perform this wait while holding the lock.

[Substrate:a37817b1572230c92ea5efbd64ed24caa45d5f63]