ESE Block Cache:  fix race on IFileSystemFilter::ErrFileOpenById that resulted in a recovery failure in accept stress
There is a race between IFileSystemFilter::ErrFileOpenById and IFileAPI::ErrRename/IFileFilter::ErrFileRename that can cause ErrFileOpenById to transiently fail with JET_errFileNotFound.  This is bad because this result can cause the block cache to conclude that a file no longer exists and thus its cached data is invalid and may be purged.

In accept stress, this race resulted in a recovery failure.  This is because a portion of a transaction log that was repeatedly reused was purged resulting in an empty segment (because the segment was empty in the underlying file as it had never been written out) followed by valid older segments.  This tripped the log's corruption detection logic which failed with 'EmptySegmentHigherLgen'.  The log file was in a state that could have been recognized as valid so this case is more of a compatibility issue than a true corruption.

[Substrate:518587b2681f60cd81323d5ccfc8e3e2596aeadd]