ESE Block Cache:  fix sparse file map tracking again
EBC tracks the sparseness of the underlying file to ensure that it does the appropriate write throughs to make a file unsparse as needed.  EBC intercepts ErrSetSize, ErrIOTrim, and ErrIOWrite to achieve this.

The problem is that ICache::ErrInvalidate is called *before* ErrSetSize actually changes the file size.  So if this fails then we may have purged a record of sparse file ranges w/o fixing the file to match.  This can cause problems later.

The fix is to pessimistically retain sparse ranges until explicitly reset via a ErrIOWrite call (or the file is closed/reopened).  This means if the file shrinks and regrows through offsets that were sparse then we will force those to be write though when in theory that isn't necessary.

[Substrate:0dba1a5928f1511283fd65dbe09e3e82f4c23235]