ESE Block Cache:  defensive changes for durability
This change attempts to further strengthen the guarantee that we are doing the correct things to maintain durability and recoverability in the cache.  These include:
-  Code reorg to ensure we only write slabs that are durable for write back
-  Changing the order in which we acquire information to drive truncation of the journal
-  Walking the entire list of dirty slabs to compute the min jpos rather than relying on the list being ordered by jpos

This change also cleans up some obsolete code related to internal opens failing due to a pending external close which is no longer possible.  (This was removed in a previous change.)

[Substrate:98796f6662366cae45dc1df75d5683aebedb9945]