ESE Flush Map header latch can deadlock the IO thread
This change fixes the following deadlock that can only happen when IFileAPI::ErrFlushFileBuffers can do async IO such as when the ESE Block Cache is enabled:

-  A FM header async write completion on the (single) IO thread is blocked waiting to upgrade the X latch to  the W latch, which blocks async IO and IOREQ production.
-  A thread is querying the FM for the min log generation via CFlushMap::LGetFmGenMinRequired.  This thread gets the S latch and calls FFB on the FM file.  This FFB must wait for pending async IO to complete

The fix is to change the FM write completion so that it can run async off of the IO thread if it would block trying to get the W latch.

This fix is far more complex than a simpler fix which would be to just change CFlushMap::LGetFmGenMinRequired to acquire the X latch rather than the S latch to avoid the above deadlock situation, at the expense of making checkpoint computation slower in rare cases.

[Substrate:c84323e39f74b6585cb257795badd4fdcb68cadb]