Fix issue causing AD database defrag to fail
Offline defrag uses parallel index rebuild which rebuilds multiple indices on multiple threads using the same PIB, storing ptlsApi on the PIB during the callback for querying callback columns results in improper sharing of TLS and the fInCallback state getting incorrectly cleared causing us to finally hit session sharing violation. So, just not cache TLS in the PIB when in a callback.

[Substrate:b0394d67d1ba794d777cb367153f35a30617603e]