ESE Block Cache:  Torn write resiliency for the Cached Block Write Counts
This change implements torn write protection for the Cached Block Write Counts by allowing N sets of write counts and rotating writes between them for every save.  When N = 2 (as it will be), this is effectively shadow paging.  This is acceptable because the size of the write counts is negligible with respect to the size of any cache.

A set of write counts is only considered valid if each chunk passes its own verification (offset, checksum) and if the entire set have the same write count and the same random number.  This allows us to detect cases where a part of a set has been lost.  It also prevents us from accidentally making a previously invalid set valid via a partial overwrite.  The write count allows us to choose the current set when N valid sets are present.

We also added a missing flush when saving the write counts.

This was the last part of EBC that is not protected from torn writes.  This was found during review.

[Substrate:18de04c4a4bb71a5313e795f03bc13fe84079add]