ESE Block Cache:  more configuration fixes
The E2E test case was trying to verify that JetGetDatabaseFileInfo worked but it failed to actually check the result.  It seems that JGDFI will return garbage if the header isn't recognized.

This change fixes the config again so that we do not cache the EBC enable state.  This state was getting cached in g_pfsconfigGlobal before the config was set by ExchangeDatabaseManager.

HA FileChecker couldn't query the state of the database copy because the EBC config was cached as off in this way.

[Substrate:71b28057b4f0b3b8bc67896baed6a6ec848b1423]