ESE Block Cache:  Fix concurrent RO file access race
The File System Filter had a state management bug where it was allowing unsynchronized access to part of a file's state.  This would only show up for files that could be open concurrently such as files open for read only access.  This isn't common but does happen.  This broke some ESE RBS tests that used this behavior.

The fix is to defer the release of any COpenFile associated with a CFilePathTableEntry until that entry is released.  Extra COpenFile instances per CFilePathTableEntry are rare (rename/move) so this shouldn't cause any problems.

We also fixed non-atomic updates of the ref count on the CFilePathTableEntry.  Lost updates on this count resulted in file filter references pointing to released files.  This is likely what caused the purecall failures.

This also fixed some stale codegen in PerseusSources.txt

[Substrate:a9a445c62c96cd2816e9f0a5e0e6f2e8cbc4082c]