ESE Block Cache:  torn writes
We are seeing a low incidence of torn writes.  These torn writes are being successfully patched by the system.

We believe that the root cause of this is a misconfiguration of the cache where the consecutive bytes of cache file held per hash slab is less than the database page size.  This results in an opportunity for a torn write on a process crash because the "transactional" unit is the set of cache blocks updated per journal entry which is bounded by a slab.

The fix is to change CachedFilePerSlab=32768.

This change also includes other related fixes:

- new instrumentation to indicate a -1018 observed by EBC vs ESE
- new instrumentation to detect missed torn write opportunities that aren't covered by this misconfiguration

[Substrate:3543f5790835ae9399069c5200024dbcbee15bbb]