ESE Block Cache:  Concurrent IO Configuration
This largely mechanical change moves the global configuration for max concurrent IO and max concurrent background IO into IFileSystemConfiguration.  The motivation is to make this configuration available to EBC so that it can participate in IO throttling.

These settings are consumed by COSDisk and are cached when the COSDisk is created.  The net result will be similar to the existing scheme.  The only difference is that, in theory, COSDisks can get different max concurrent background IO configurations over time because we do not block setting JET_paramCheckpointIOMax after global init.  I did not fix this because I suspect it would cause an app compat problem.

There were a few truly global uses of g_cioOutstandingMax that had to be addressed.

[Substrate:0d36209fdd6b2160a83908f57014c876fd9d5dfa]