ESE Block Cache:  Clean Slab Visitor fixes
During stress, we hit a case where we generated a torn write opportunity due to cleaning some but not all of the space we needed.  A subsequent update was able to clean more space due to concurrent release of IO range locks, enabling us to hit this rather than a simple clean failure.

When investigating this, I noticed that if you do a clean to cache a 32kb page then we will invoke clean for every 4kb of the 32kb block (i.e. 8x).  That sucks.

This change fixes the clean process to use local state to remember how much space it produced recently so that it doesn't have to clean more than necessary.  I tried to do something like this before but got stuck because I couldn't compute fChangedSlab properly.  I realized why.  The pcbs is actually a wrapper and changing slabs caused us to release and reallocate the same wrapper immediately, invalidating any attempt to check for a change in pcbs to indicate a change in slab.  We must return a flag set when we actually open a new slab.  Existing code to check for a torn write opportunity had the same bug and was fixed.

[Substrate:0d6eeb615cdcb9f09ca552151a4a4c5699ac932f]