ESE Block Cache:  IO Throttling
This change removes the existing partial short circuit of IO throttling for cached files and provides another partial implementation of IO throttling for cached files in TFileFilter.  The portions of IO throttling that are provided are related to max IO counts and IO combining.  This does not change how IO scheduling occurs (e.g. low priority IO, IO smoothing) because this must happen in the IO queue itself.

We must implement IO throttling at this level to allow proper throttling of engine IO as it expects while allowing the cache to actually perform a subset of this IO underneath.  This subset of IO to the cached file (cache misses and cache writethrough) are not throttled because they are executed on behalf of the original throttled requests.  Caching file IO may also be performed.  This IO explicitly bypasses the max concurrent IO check.  We do this to prevent deadlocks caused by one IO consuming more than one throttle count because there is no way to release the original IO throttle count without performing the caching file IO to complete that request.

The IO throttling is implemented by TFileFilter as a mathematical model only.  It only tracks enough IO state per thread to compute IO combining and its impact on IO throttling.  It only tracks enough IO state to compute the IO queues per scheduling context (IFileFilter::ErrDiskId).  When an IO count is consumed by the first non-combinable IO, this fact is remembered by the relevant IO completion context and the IO count is released when that IO completes.  When in "IOREQ" is reserved via IFileAPI::ErrReserveIOREQ, a context is returned that may also capture this same state.  This state is transferred to a real IO completion context when the "IOREQ" is "used" by a read or it is released by IFileAPI::ReleaseUnusedIOREQ.

The following supporting changes were made:

-  Some code in BF that was written to assume a single IO thread has been changed to work during an IO completion.  This is required because EBC completes IOs on any number of thread pool threads.
-  Some minor changes to ErrIOWriteContiguous were made.
-  Some functionality in OSDisk that is used to setup IO throttling limits was exposed internally in the OS layer for reuse by TFileFilter.
-  A cleanup method was added to CacheThreadLocalStorage to facilitate its use in a static global.
-  We cleaned up how preallocated IOREQs are released.

[Substrate:fcec8b4baba7855f87705dcb0f4f086260a6b46b]